Add program to print the name of the program
This commit adds a new file, `0-whatsmyname.c`, which contains a program that prints the name of the program. The program takes the number of arguments and an array of arguments as input and prints the first argument, which is the name of the program.

New files:
- `0-whatsmyname.c`: Contains the program to print the name of the program

Modified files:
None